at86rf215: get rid of at86rf215_set_state()
at86rf215_set_state() was a misguided cargo-cult takeover from at86rf2xx.
The state-machine of at86rf215 is much simpler and at86rf215_set_state()
with it's 'any-to-any' state approach turned out to be dangerous because it
works independent of the driver's own state machine.

Remove it and replace it by some finer grained functions.

Also turnes out we don't have to switch to TRXOFF to switch radio modes.